🧹 Fix unused import code health issue in drilldown.ts
