+++ packages/drilldown/src/cli.ts
@@ -4,7 +4,8 @@ import { parsePositiveInt } from "@paper-tools/core";
 import "dotenv/config";
 import { Command } from "commander";
 import { searchByKeyword, searchByVenue, enrichAllWithCrossref, searchCrossref } from "./search.js";
-import { drilldown, extractKeywords } from "./drilldown.js";
+import { extractKeywords } from "./drilldown.js";
+import { drilldown } from "./search.js";
 import { readFileSync } from "node:fs";
 import { fileURLToPath } from "node:url";
 import { dirname, join } from "node:path";
