+++ packages/drilldown/src/search.ts
@@ -1,3 +1,4 @@
+import { extractKeywords } from "./drilldown.js";
 import type { Paper } from "@paper-tools/core";
 import {
     searchPublications,
@@ -92,4 +93,94 @@ export async function searchCrossref(
     maxResults = 20,
 ): Promise<Paper[]> {
     return searchWorks(query, maxResults);
+}
+
+/**
+ * drilldown 結果の型
+ */
+export interface DrilldownResult {
+    /** 探索の深さレベル (0 = seed) */
+    level: number;
+    /** この深さで見つかった論文 */
+    papers: Paper[];
+}
+
+/**
+ * シード論文群から再帰的にキーワード抽出 → 検索を繰り返して深掘りする
+ * @param seedPapers - 初期論文リスト
+ * @param depth - 深掘りの最大深さ（デフォルト 1）
+ * @param maxPerLevel - 各レベルで取得する最大論文数（デフォルト 10）
+ * @param enrich - Crossref で情報を補完するか（デフォルト false）
+ */
+export async function drilldown(
+    seedPapers: Paper[],
+    depth = 1,
+    maxPerLevel = 10,
+    enrich = false,
+): Promise<DrilldownResult[]> {
+    const results: DrilldownResult[] = [{ level: 0, papers: seedPapers }];
+    const seenDois = new Set<string>();
+    const seenTitles = new Set<string>();
+
+    // seed の DOI と title を記録
+    for (const p of seedPapers) {
+        if (p.doi) seenDois.add(p.doi.toLowerCase());
+        if (p.title) {
+            const normalizedTitle = p.title.toLowerCase().trim().replace(/\s+/g, " ");
+            seenTitles.add(normalizedTitle);
+        }
+    }
+
+    let currentPapers = seedPapers;
+    const enrichPromises: Promise<void>[] = [];
+
+    for (let d = 1; d <= depth; d++) {
+        const keywords = extractKeywords(currentPapers, 5);
+        if (keywords.length === 0) break;
+
+        const query = keywords.join(" ");
+        let found = await searchByKeyword(query, maxPerLevel * 2);
+
+        // 既出 DOI / title を除外
+        found = found.filter((p) => {
+            if (p.doi) {
+                const lower = p.doi.toLowerCase();
+                if (seenDois.has(lower)) return false;
+                seenDois.add(lower);
+            } else if (p.title) {
+                const normalizedTitle = p.title.toLowerCase().trim().replace(/\s+/g, " ");
+                if (seenTitles.has(normalizedTitle)) return false;
+                seenTitles.add(normalizedTitle);
+            }
+            return true;
+        });
+
+        found = found.slice(0, maxPerLevel);
+
+        if (found.length === 0) break;
+
+        const resultEntry: DrilldownResult = { level: d, papers: found };
+        results.push(resultEntry);
+
+        if (enrich) {
+            // Background enrichment allows parallel fetch
+            enrichPromises.push(
+                enrichAllWithCrossref(found)
+                    .then((enriched) => {
+                        resultEntry.papers = enriched;
+                    })
+                    .catch((error) => {
+                        const message = error instanceof Error ? error.message : String(error);
+                        console.error(`[drilldown] Enrichment failed at level ${d}: ${message}`);
+                    })
+            );
+        }
+        currentPapers = found;
+    }
+
+    if (enrichPromises.length > 0) {
+        await Promise.all(enrichPromises);
+    }
+
+    return results;
 }
